pyproject.toml
[project]
name = "mkapartner-ru"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-admin-sortable2 (>=2.2.6,<3.0.0)",
    "sorl-thumbnail (>=12.11.0,<13.0.0)",
    "transliterate (>=1.10.2,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[tool.poetry]
packages = [{include = "mkapartner"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
